Enable mentored bugs from bugzilla to be shown
-Refactor project/index.jsx
-Separate it into two routes, /projects/:project/mentored and
/projects/:project
-Move the AppBar into the project/index.jsx since it is similar
for both route
-Add mentored bugs filter using f1 and o1 properties

Closes #32
